requirements.txt
opencv-python
mediapipe
pandas
tqdm
numpy
scikit-learn
tensorflow
flask
pillow
werkzeug
matplotlib
seaborn 